Refactor API and frontend workflow helpers for maintainability
